:beetle: html-proofer: Remove error-sort flag to restore error verbosity
This commit removes the `error-sort` flag from the `htmlproofer` command
in order to restore logs from HTML Proofer in the Circle CI build
report.

The current behavior with this flag added is to suppress all error
messages, which is not helpful when you have a failing CI pipeline from
HTML Proofer. To debug requires a local checkout of the repo and to
manually run HTML Proofer to see the errors.

I tested this change locally and verified that removing this flag
restores the errors in log messages. This is an uncommon flag to use for
HTML Proofer, and I think it does not do exactly what it is expected to
do, so the best option appears to be removing it, thus simplifying what
needs to be maintained here.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>